LPD-58072 Merge StatsIndexRequestExecutor/StatsIndexRequestExecutorImpl into pojos
